scripts/requirements.txt
scholarly
pyyaml
